backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6  # For file uploads
python-dotenv>=1.0.0     # For environment variables

# LangChain and related dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-chroma>=0.0.10
chromadb>=0.4.22        # Vector store

# Embedding models
sentence-transformers>=2.2.2
torch>=2.1.0

# Document processing
unstructured>=0.11.0    # For URL content extraction
python-docx>=1.0.1      # For DOCX files
python-pptx>=0.6.21     # For PPT files
PyPDF2>=3.0.0          # For PDF files
beautifulsoup4>=4.12.0  # For web scraping
requests>=2.31.0        # For HTTP requests